+++ src/api-contract/types.ts
@@ -224,3 +224,29 @@ export type ShlinkShortUrlsListParams = {
   excludeMaxVisitsReached?: boolean;
   excludePastValidUntil?: boolean;
 };
+
+export type ShlinkRedirectConditionType = 'device' | 'language' | 'query-param';
+
+export type ShlinkRedirectCondition = {
+  type: ShlinkRedirectConditionType;
+  matchKey: string | null;
+  matchValue: string;
+};
+
+export type ShlinkRedirectRuleData = {
+  longUrl: string;
+  conditions: ShlinkRedirectCondition[];
+};
+
+export type ShlinkRedirectRule = ShlinkRedirectRuleData & {
+  priority: number;
+};
+
+export type ShlinkRedirectRulesList = {
+  defaultLongUrl: string;
+  redirectRules: ShlinkRedirectRule[];
+};
+
+export type ShlinkSetRedirectRulesData = {
+  redirectRules: ShlinkRedirectRuleData[];
+};