+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 * [#72](https://github.com/shlinkio/shlink-js-sdk/issues/72) Add support for Shlink 4.0.0
 
+  * Add support to list and set short URL redirect conditions.
   * Add `type` param to `ShlinkApiClient.getOrphanVisits`.
 
 ### Changed